[WIP] UI refactor

+++ service/src/main/java/com/github/kr328/clash/service/data/ClashProfileProxyDao.kt
@@ -11,7 +11,7 @@ interface ClashProfileProxyDao {
     suspend fun setSelectedForProfile(item: ClashProfileProxyEntity)
 
     @Query("SELECT * FROM profile_select_proxies WHERE profile_id = :id")
-    suspend fun querySelectedForProfile(id: Int): List<ClashProfileProxyEntity>
+    suspend fun querySelectedForProfile(id: Long): List<ClashProfileProxyEntity>
 
     @Query("DELETE FROM profile_select_proxies WHERE profile_id = :id AND proxy in (:selected)")
     suspend fun removeSelectedForProfile(id: Int, selected: List<String>)
