+++ service/src/main/java/com/github/kr328/clash/service/data/ClashProfileProxyEntity.kt
@@ -14,7 +14,7 @@ import androidx.room.*
     )]
 )
 data class ClashProfileProxyEntity(
-    @ColumnInfo(name = "profile_id") val profileId: Int,
+    @ColumnInfo(name = "profile_id") val profileId: Long,
     @ColumnInfo(name = "proxy") val proxy: String,
     @ColumnInfo(name = "selected") val selected: String,
     @ColumnInfo(name = "id") @PrimaryKey(autoGenerate = true) val id: Int = 0
